Fixed broken tests caused by changing button names.
